chore: Remake example project in Xcode 16
