+++ Example/PokemonAPI-App/Assets.xcassets/AccentColor.colorset/Contents.json
