+++ Example/PokemonAPI-App/Assets.xcassets/Contents.json
