+++ Example/PokemonAPI-App/Image+.swift
