+++ Example/PokemonAPI-App/PokemonAPI+.swift
@@ -1,0 +1,12 @@
+//
+//  PokemonAPI+.swift
+//  PokemonAPI-App
+//
+//  Created by Christopher Jennewein on 11/17/24.
+//
+
+import Foundation
+import PokemonAPI
+
+
+extension PokemonAPI: @retroactive ObservableObject { }
