--- Example/PokemonAPI-App/Utilities/PokemonAPI+.swift
@@ -1,12 +1,0 @@
-//
-//  PokemonAPI+.swift
-//  PokemonAPI-App
-//
-//  Created by Christopher Jennewein on 10/3/21.
-//
-
-import Foundation
-import PokemonAPI
-
-
-extension PokemonAPI: ObservableObject { }
